requirements.txt
streamlit
langchain
langchain_deepseek
langchain_chroma
langchain_text_splitters
chromadb
deepseek-ai
python-dotenv
# langchain_openai  # Removed - not using OpenAI embeddings anymore
streamlit_agraph
langgraph
langchain-community
pypdf
sentence-transformers  # *** IMPORTANT: ADDED sentence-transformers - for HuggingFace embeddings ***

langchain-huggingface  # Updated dependency